Implement adding comments

+++ app/controllers/articles_controller.rb
@@ -24,6 +24,9 @@ class ArticlesController < ApplicationController
   end
 
   def show 
+      @comment = @article.comments.build
+      @comments = @article.comments
+
   end
 
   def edit
@@ -49,9 +52,14 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
-    if @article.destroy
+    unless @article.user == current_user
+      flash[:alert] = "You can only delete your own article."
+      redirect_to root_path
+    else
+       if @article.destroy
         flash[:success] = "Article has been deleted"
         redirect_to articles_path
+      end
     end
   end
